+++ lib/object/index.ts
@@ -1,0 +1,19 @@
+'use strict';
+
+import {define}             from './define';
+import {isObject}           from './is';
+import {isNotEmptyObject}   from './isNotEmpty';
+import {merge}              from './merge';
+import {pick}               from './pick';
+
+export {
+    define,
+    isObject,
+    isObject as is,
+    isNotEmptyObject,
+    isNotEmptyObject as isNotEmpty,
+    isNotEmptyObject as isNeObject,
+    isNotEmptyObject as isNe,
+    merge,
+    pick,
+};