+++ CHANGELOG.md
@@ -21,6 +21,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - **feat**: hash/* namespace export
 - **feat**: number/* namespace export
 - **feat**: number namespace export has the following shorthand aliases: `is`, `isGt`, `isGte`, `isLt`, `isLte`, `isNum`, `isNumGt`, `isNumGte`, `isNumLt`, `isNumLte`, `isInt`, `isIntGt`, `isIntGte`, `isIntLt`, `isIntLte`, `isIntBetween`, `randBetween`, `randIntBetween`, `toPct`
+- **feat**: object/* namespace export
+- **feat**: object namespace export has the following shorthand aliases: `is`, `isNeObject`, `isNotEmpty`, `isNe`
 - **feat**: formdata/is - Utility to verify whether or not a passed value is an instance of FormData, also acts as a typeguard
 - **feat**: formdata/* namespace export
 - **feat**: formdata namespace export has the following shorthand aliases: `is`